+++ build.sh
@@ -7,14 +7,9 @@ if ! command -v xcodebuild >/dev/null 2>&1; then
     exit 0
 fi
 
-ci_build() {
-    NAME=$1
-    xcodebuild -project "touchid.xcodeproj" \
-               -target "touchid" \
-               -destination "platform=iOS Simulator,name=${NAME}" \
-               -sdk iphonesimulator \
-               -configuration "Debug" \
-               build
-}
-
-ci_build "${SIMULATOR_NAME:-iPhone 6}"
+xcodebuild -project "touchid.xcodeproj" \
+           -target "touchidTests" \
+           -sdk iphonesimulator \
+           -configuration "Debug" \
+           CODE_SIGNING_ALLOWED=NO \
+           build